Product app: Add delete confirmation functionality and template

+++ products/views.py
@@ -143,6 +143,14 @@ def delete_product(request, product_id):
         return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.success(request, 'Product deleted!')
-    return redirect(reverse('products'))
+    if request.method == 'POST':
+        # Confirm delete
+        product.delete()
+        messages.success(request, 'Product deleted!')
+        return redirect(reverse('products'))
+
+    context = {
+        'product': product,
+    }
+
+    return render(request, 'products/delete_product.html', context)
